Deserialize / validate default values
