+++ umongo/abstract.py
@@ -140,6 +140,18 @@ class BaseField(ma_fields.Field):
 
         super().__init__(*args, **kwargs)
 
+        # Deserialize default/missing values
+        # This ensures they are validated and get the proper types and constraints
+        for attr in ('default', 'missing'):
+            default = getattr(self, attr)
+            if default is not missing:
+                if callable(default):
+                    def call_default():
+                        return self.deserialize(default())
+                    setattr(self, attr, call_default)
+                else:
+                    setattr(self, attr, self.deserialize(default))
+
         # Overwrite error_messages to handle i18n translation
         self.error_messages = I18nErrorDict(self.error_messages)
         # `io_validate` will be run after `io_validate_resursive`
