+++ umongo/fields.py
@@ -53,18 +53,6 @@ __all__ = (
 
 class DictField(BaseField, ma_fields.Dict):
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        for attr in ('default', 'missing'):
-            default = getattr(self, attr)
-            if default is not missing:
-                if callable(default):
-                    def call_default():
-                        return Dict(default())
-                    setattr(self, attr, call_default)
-                else:
-                    setattr(self, attr, Dict(default))
-
     def _deserialize(self, value, attr, data):
         value = super()._deserialize(value, attr, data)
         return Dict(value)
@@ -86,18 +74,6 @@ class DictField(BaseField, ma_fields.Dict):
 
 
 class ListField(BaseField, ma_fields.List):
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        for attr in ('default', 'missing'):
-            default = getattr(self, attr)
-            if default is not missing:
-                if callable(default):
-                    def call_default():
-                        return List(self.container, default())
-                    setattr(self, attr, call_default)
-                else:
-                    setattr(self, attr, List(self.container, default))
 
     def _deserialize(self, value, attr, data):
         return List(self.container, super()._deserialize(value, attr, data))
